change singlular member controller back to plural, add some database defaults for members, and add some authentication checks for member route access
