+++ routes/web.php
@@ -11,15 +11,15 @@
 |
 */
 
-Route::get('/', 'MemberController@index')->name('home');
-Route::get('/member/create', 'MemberController@create');
-Route::post('/member', 'MemberController@store');
-Route::get('/member/{member}', 'MemberController@show');
-Route::get('/member/edit/{member}', 'MemberController@edit');
+Route::get('/', 'MembersController@index')->name('home');
+Route::get('/member/create', 'MembersController@create');
+Route::post('/member', 'MembersController@store');
+Route::get('/member/{member}', 'MembersController@show');
+Route::get('/member/edit/{member}', 'MembersController@edit');
 
 Route::get('/user/create', 'UserController@create');
 Route::post('/user', 'UserController@store');
-Route::get('/user/{user}', 'UserController@show')->name('user');
+Route::get('/user/{user}', 'UserController@show');
 
 Route::get('/login', 'SessionsController@create')->name('login');
 Route::post('/login', 'SessionsController@store');
